BF: Using uint64_t causes problems on 32bit platforms

+++ indexed_gzip/zran.h
@@ -165,11 +165,8 @@ struct _zran_index {
      * against the CRC and size in the gzip
      * footer, and an error is returned if
      * they don't match.
-     *
-     * (the zlib crc32 function returns
-     * an unsigned long, so we use u64)
      */
-    uint64_t stream_crc32;
+    uint32_t stream_crc32;
     uint32_t stream_size;
     uint8_t  validating;
 };
